refactor: reorganize controller packages and update FXML references for flock and supply management
