--- app/src/main/kotlin/org/db_poultry/controller/GenerateReportController.kt
@@ -1,20 +1,0 @@
-package org.db_poultry.controller
-
-import org.db_poultry.util.SceneSwitcher
-import org.db_poultry.util.GeneralUtil
-
-import javafx.fxml.FXML
-import javafx.scene.text.Text
-
-class GenerateReportController {
-
-    @FXML
-    private lateinit var flockStartDateLbl: Text
-
-    @FXML
-    private lateinit var mortalityRateLbl: Text
-
-    @FXML
-    private lateinit var qtyStartedLbl: Text
-
-}
