+++ app/src/main/kotlin/org/db_poultry/controller/flock/CreateFlockMenuController.kt
@@ -1,4 +1,4 @@
-package org.db_poultry.controller
+package org.db_poultry.controller.flock
 
 import org.db_poultry.util.GeneralUtil
