+++ app/src/main/kotlin/org/db_poultry/controller/flock/FlockGridHomeController.kt
@@ -1,4 +1,4 @@
-package org.db_poultry.controller
+package org.db_poultry.controller.flock
 
 import org.db_poultry.util.GeneralUtil
 
@@ -14,9 +14,10 @@ import org.db_poultry.db.flockDAO.ReadFlock
 import org.db_poultry.pojo.FlockPOJO.FlockComplete
 import java.net.URL
 import java.util.ResourceBundle
+import kotlin.collections.iterator
 
 // TODO: Revise naming of this class to better reflect its purpose @Dattebayo2505
-class ViewController: Initializable {
+class FlockGridHomeController: Initializable {
 
     @FXML
     lateinit var selectFlockAnchorPane: AnchorPane
