+++ app/src/main/kotlin/org/db_poultry/controller/flock/popup/FlockCreateNewController.kt
@@ -1,10 +1,4 @@
-package org.db_poultry.controller
-
-import org.db_poultry.db.DBConnect.getConnection
-import org.db_poultry.db.flockDAO.CreateFlock.createFlock
-import org.db_poultry.util.GeneralUtil
-import org.db_poultry.util.undoSingleton
-import org.db_poultry.util.undoTypes
+package org.db_poultry.controller.flock.popup
 
 import javafx.fxml.FXML
 import javafx.scene.control.Button
@@ -13,11 +7,15 @@ import javafx.scene.control.TextField
 import javafx.scene.layout.AnchorPane
 import javafx.scene.shape.Rectangle
 import javafx.scene.text.Text
+import org.db_poultry.db.DBConnect
+import org.db_poultry.db.flockDAO.CreateFlock
 import org.db_poultry.util.PopupUtil
+import org.db_poultry.util.undoSingleton
+import org.db_poultry.util.undoTypes
 import java.sql.Date
 
-class CreateNewFlockController {
-    // private var jdbcURL: String    
+class FlockCreateNewController {
+    // private var jdbcURL: String
 
     @FXML
     private lateinit var anchorPane: AnchorPane
@@ -53,7 +51,7 @@ class CreateNewFlockController {
         println("\nStart Count: $startCount")
         println("Date: $date")
 
-        if (createFlock(getConnection(), startCount, date) != null) {
+        if (CreateFlock.createFlock(DBConnect.getConnection(), startCount, date) != null) {
             undoSingleton.setUndoMode(undoTypes.doUndoFlock)
             PopupUtil.showPopup("success", "Flock created successfully!")
             println("Successfully created Flock.")
